+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Version 0.6.4 (TBA)
+* Improvement: Replaced ArrayList by thread-safe variant in WebsocketClient to speed-up message processing
 
 # Version 0.6.3 (14.04.2018)
 * New Feature: Added the connection feature manager